pyproject.toml
[tool.poetry]
name = "llvm-exp"
version = "0.1.0"
description = ""
authors = ["Santiago SILVA <16252054+sssilvar@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.0.323"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
pandas = "^2.1.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
